Remove extra header from confirmation pages in expanded view
