+++ ui/pages/confirmations/components/confirm-page-container/confirm-page-container-header/confirm-page-container-header.component.js
@@ -64,7 +64,7 @@ export default function ConfirmPageContainerHeader({
             </span>
           </div>
         )}
-        {isFullScreen ? null : <NetworkDisplay />}
+        <NetworkDisplay />
       </div>
       {children}
     </div>
