+++ ui/pages/confirmations/hooks/useConfirmationRecipientInfo.ts
@@ -11,11 +11,11 @@ function useConfirmationRecipientInfo() {
   const currentConfirmation = useSelector(
     currentConfirmationSelector,
   ) as Confirmation;
+  const allAccounts = useSelector(accountsWithSendEtherInfoSelector);
 
   let senderAddress, senderName;
   if (currentConfirmation) {
     const { from } = getConfirmationSender(currentConfirmation);
-    const allAccounts = useSelector(accountsWithSendEtherInfoSelector);
     const fromAccount = getAccountByAddress(allAccounts, from);
 
     senderAddress = from;
